Foundations of Chantry Cardinal System
The Chantry has several specific Grand Clerics that elect new Divines. Orlesian Chantry I took the ones we knew, added the number of Clerics as seen in 'Dawn of the Seeker' and their colours, and extrapolated the ones that were missing.

The Imperial Chantry ran a similar system from more local Clerics, but I decided to add in exiled provincial Clerics as well. Either non existent and added in if Tevinter reclaims portions of Thedas, or use them as small prestige titles to be appropriated by aspiring Tevinter nobles. (After all, some of these titles would have been meaningless for nearly 1000 years.) This system allows Tevinter to have lots of voting Grand Clerics, but not flooding every city with them.